+++ server/helpers/dbImport/posts.ts
@@ -1,17 +1,18 @@
 import { PoolClient } from "pg";
 import { Database } from "better-sqlite3";
 import { Import } from "./import";
+import { Service } from "./index";
 
 export default class Posts extends Import {
   display = "Posts";
   
-  service = 0;
-  inputTable = () => `current_files_${this.service}`;
+  service?: Service;
+  inputTable = () => `current_files_${this.service?.id || 0}`;
   
   batchSizeMul = 1 / 2;
   outputTable = "posts";
   totalQuery = () => `SELECT count(1) FROM ${this.inputTable()}`;
-  outputQuery = (table: string) => `COPY ${table}(id, sha256, md5, blurhash, size, width, height, duration, num_frames, has_audio, rating, mime, posted) FROM STDIN (FORMAT CSV)`;
+  outputQuery = (table: string) => `COPY ${table}(id, sha256, md5, blurhash, size, width, height, duration, num_frames, has_audio, rating, mime, inbox, trash, posted) FROM STDIN (FORMAT CSV)`;
   inputQuery = () => `
     SELECT
       current_files.hash_id,
@@ -27,6 +28,8 @@ export default class Posts extends Import {
       COALESCE(files_info.has_audio, '') || ',' ||
       COALESCE(local_ratings.rating, '') || ',' ||
       COALESCE(files_info.mime, '') || ',' ||
+      (file_inbox.hash_id IS NOT NULL) || ',' ||
+      ${this.service?.trash || false} || ',' ||
       datetime(current_files.timestamp, 'unixepoch', 'utc') || '\n'
     FROM ${this.inputTable()} current_files
       INNER JOIN hashes ON hashes.hash_id = current_files.hash_id
@@ -34,6 +37,7 @@ export default class Posts extends Import {
       LEFT JOIN local_hashes ON local_hashes.hash_id = current_files.hash_id
       LEFT JOIN blurhashes ON blurhashes.hash_id = current_files.hash_id
       LEFT JOIN local_ratings ON local_ratings.service_id = ${this.ratingService} AND local_ratings.hash_id = current_files.hash_id
+      LEFT JOIN file_inbox ON file_inbox.hash_id = current_files.hash_id
     WHERE current_files.hash_id > ?
     ORDER BY current_files.hash_id ASC
     LIMIT ?
@@ -43,7 +47,7 @@ export default class Posts extends Import {
     super(hydrus, postgres);
   }
   
-  async startEach(services: number[]) {
+  async startEach(services: Service[]) {
     const sizes = services.map(service => {
       this.resetTotal();
       this.service = service;