+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 # Unreleased
 
+- Added support for `system:inbox`, `system:archive` and `system:trash` (https://github.com/funmaker/Hybooru/issues/25)
 - Added `host` config option (https://github.com/funmaker/Hybooru/issues/38)
 - Added support for PORT and HOST envvar override
 - Removed DOCKERIZED envvar, Dockerfile now uses PORT (https://github.com/funmaker/Hybooru/issues/36)